!3 组件升级spring boot->2.1.1 mybatis plus-> 3.0.7.1 Merge pull request !3 from tankxxl/master

+++ kernel-model/src/main/java/cn/stylefeng/roses/kernel/model/page/PageResult.java
@@ -15,7 +15,7 @@
  */
 package cn.stylefeng.roses.kernel.model.page;
 
-import com.baomidou.mybatisplus.plugins.Page;
+import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import lombok.Data;
 
 import java.io.Serializable;
@@ -44,8 +44,8 @@ public class PageResult<T> implements Serializable {
     public PageResult(Page<T> page) {
         this.setRows(page.getRecords());
         this.setTotalRows(page.getTotal());
-        this.setPage(page.getCurrent());
-        this.setPageSize(page.getSize());
+        this.setPage((int) page.getCurrent());
+        this.setPageSize((int) page.getSize());
     }
 
 }
